iris-vector-db/requirements.txt
sqlalchemy-iris
langchain-iris
testcontainers-iris
llama-iris
llama-index-legacy
sentence-transformers
langchain
fastembed 
openai 
tiktoken
python-dotenv
pandas
ipykernel
setuptools
fastapi
pyngrok
uvicorn